[openstack] Support for offline resize
